+++ dotnet/src/ComponentDiscovery.DspAnalysis.net48/My Project/AssemblyInfo.vb
@@ -16,6 +16,6 @@ Imports System.Runtime.InteropServices
 <Assembly: ComVisible(False)>
 <Assembly: Guid("5027f65d-44ad-44b7-805a-2de61c83390e")>
 
-<Assembly: AssemblyVersion("4.10.3")>
-<Assembly: AssemblyInformationalVersion("4.10.3")>
-'WARNING: DONT SPECIFY: <Assembly: AssemblyFileVersion("4.10.3")> 
+<Assembly: AssemblyVersion("4.11.0")>
+<Assembly: AssemblyInformationalVersion("4.11.0")>
+'WARNING: DONT SPECIFY: <Assembly: AssemblyFileVersion("4.11.0")>