+++ doc/changelog.md
@@ -8,9 +8,14 @@ This files contains a version history including all changes relevant for semanti
 
 ## Upcoming Changes
 
-* **new Feature**: added '*InstanceDiscoveryContext.UseExternalManagedTypeIndexer*' and '*InstanceDiscoveryContext.ActivateScoping*' to allow less invasive customizing of runtime scopes
+*(none)*
 
-* Fix: not more NullReferenceException when InstanceDiscovery reading provided instances from Private Getter-Methods
+
+
+## v 4.11.0
+released **2025-12-10**, including:
+ - **new Feature**: added '*InstanceDiscoveryContext.UseExternalManagedTypeIndexer*' and '*InstanceDiscoveryContext.ActivateScoping*' to allow less invasive customizing of runtime scopes
+ - Fix: not more NullReferenceException when InstanceDiscovery reading provided instances from Private Getter-Methods